eas/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary

drf-yasg
djangorestframework-simplejwt
django-filter

django-extensions
openpyxl 
xlsxwriter 
pandas 
reportlab 
weasyprint